+++ MemfaultPackages/bort/src/main/java/com/memfault/bort/AppComponents.kt
@@ -41,6 +41,8 @@ import com.memfault.bort.requester.BugReportRequester
 import com.memfault.bort.requester.LogcatCollectionRequester
 import com.memfault.bort.requester.MetricsCollectionRequester
 import com.memfault.bort.requester.PeriodicWorkRequester
+import com.memfault.bort.requester.UptimeTickRequester
+import com.memfault.bort.requester.UptimeTickTask
 import com.memfault.bort.settings.BortEnabledProvider
 import com.memfault.bort.settings.ConfigValue
 import com.memfault.bort.settings.DynamicSettingsProvider
@@ -57,6 +59,7 @@ import com.memfault.bort.shared.PreferenceKeyProvider
 import com.memfault.bort.time.BaseAbsoluteTime
 import com.memfault.bort.time.RealBootRelativeTimeProvider
 import com.memfault.bort.time.RealCombinedTimeProvider
+import com.memfault.bort.time.UptimeTracker
 import com.memfault.bort.tokenbucket.RealTokenBucketFactory
 import com.memfault.bort.tokenbucket.TokenBucketStore
 import com.memfault.bort.tokenbucket.TokenBucketStoreRegistry
@@ -108,6 +111,7 @@ data class AppComponents(
     val dropBoxProcessedEntryCursorProvider: ProcessedEntryCursorProvider,
     val bortSystemCapabilities: BortSystemCapabilities,
     val metrics: BuiltinMetricsStore,
+    val uptimeTracker: UptimeTracker,
 ) {
     open class Builder(
         private val context: Context,
@@ -225,7 +229,9 @@ data class AppComponents(
                     .fileUploadHoldingAreaSettings::maxStoredEventsOfInterest,
             )
 
-            val tokenBucketStoreRegistry = TokenBucketStoreRegistry()
+            val uptimeTracker = UptimeTracker(sharedPreferences, readLinuxBootId())
+
+            val tokenBucketStoreRegistry = TokenBucketStoreRegistry(uptimeTracker)
             val bugReportRequestsTokenBucketStore = tokenBucketStoreRegistry.createAndRegisterStore(
                 context, "bug_report_requests"
             ) { storage ->
@@ -358,6 +364,7 @@ data class AppComponents(
                 MetricsCollectionRequester(context, settingsProvider.metricsSettings, bortSystemCapabilities),
                 BugReportRequester(context, settingsProvider.bugReportSettings),
                 LogcatCollectionRequester(context, settingsProvider.logcatSettings, bortSystemCapabilities),
+                UptimeTickRequester(context, bortSystemCapabilities),
             )
 
             val dropBoxProcessedEntryCursorProvider = ProcessedEntryCursorProvider(
@@ -396,7 +403,7 @@ data class AppComponents(
                             getMaxBuckets = { 1 },
                             getTokenBucketFactory = {
                                 RealTokenBucketFactory(
-                                    defaultCapacity = 2,
+                                    defaultCapacity = 3,
                                     defaultPeriod = settingsProvider.bugReportSettings.requestInterval *
                                         settingsProvider.bugReportSettings.periodicRateLimitingPercentOfPeriod / 100,
                                 )
@@ -465,7 +472,8 @@ data class AppComponents(
                     val millisAgo = maxOf(0, System.currentTimeMillis() - absoluteTime.timestamp.toEpochMilli())
                     val elapsedRealtime = SystemClock.elapsedRealtime() - millisAgo
                     fileUploadHoldingArea.handleEventOfInterest(elapsedRealtime.milliseconds)
-                }
+                },
+                bortSystemCapabilities = bortSystemCapabilities,
             )
 
             val httpTaskCallFactory = HttpTaskCallFactory.fromContextAndConstraints(
@@ -499,6 +507,7 @@ data class AppComponents(
                 dropBoxProcessedEntryCursorProvider = dropBoxProcessedEntryCursorProvider,
                 bortSystemCapabilities = bortSystemCapabilities,
                 metrics = builtinMetricsStore,
+                uptimeTracker = uptimeTracker,
             )
         }
     }
@@ -519,6 +528,8 @@ interface InterceptingWorkerFactory {
         reporterServiceConnector: ReporterServiceConnector,
         pendingBugReportRequestAccessor: PendingBugReportRequestAccessor,
         bugReportPeriodicTaskTokenBucketStore: TokenBucketStore,
+        bortSystemCapabilities: BortSystemCapabilities,
+        builtInMetricsStore: BuiltinMetricsStore,
     ): ListenableWorker?
 }
 
@@ -553,6 +564,7 @@ class DefaultWorkerFactory(
     private val dumpsterClient: DumpsterClient,
     private val interceptingFactory: InterceptingWorkerFactory? = null,
     private val handleEventOfInterestAtAbsoluteTime: (BaseAbsoluteTime) -> Unit,
+    private val bortSystemCapabilities: BortSystemCapabilities,
 ) : WorkerFactory(), TaskFactory {
     override fun createWorker(
         appContext: Context,
@@ -567,6 +579,8 @@ class DefaultWorkerFactory(
             reporterServiceConnector,
             pendingBugReportRequestAccessor,
             bugReportPeriodicTaskTokenBucketStore,
+            bortSystemCapabilities,
+            builtinMetricsStore,
         )?.let {
             return it
         }
@@ -584,6 +598,8 @@ class DefaultWorkerFactory(
                 pendingBugReportRequestAccessor = pendingBugReportRequestAccessor,
                 tokenBucketStore = bugReportPeriodicTaskTokenBucketStore,
                 bugReportSettings = settingsProvider.bugReportSettings,
+                bortSystemCapabilities = bortSystemCapabilities,
+                builtInMetricsStore = builtinMetricsStore,
             )
             else -> null
         }
@@ -669,6 +685,7 @@ class DefaultWorkerFactory(
                 getMaxAttempts = { 1 },
                 periodicWorkRequesters = periodicWorkRequesters,
             )
+            UptimeTickTask::class.qualifiedName -> UptimeTickTask()
             else -> null
         }
     }