+++ MemfaultPackages/bort/src/main/java/com/memfault/bort/receivers/ControlReceiver.kt
@@ -53,7 +53,15 @@ abstract class BaseControlReceiver : FilteringReceiver(
         val timeout = intent.extras?.getLongOrNull(
             INTENT_EXTRA_BUG_REPORT_REQUEST_TIMEOUT_MS
         )?.let(Long::milliseconds) ?: BugReportRequestTimeoutTask.DEFAULT_TIMEOUT
-        requestBugReport(context, pendingBugReportRequestAccessor, request, timeout, settingsProvider.bugReportSettings)
+        requestBugReport(
+            context,
+            pendingBugReportRequestAccessor,
+            request,
+            timeout,
+            settingsProvider.bugReportSettings,
+            bortSystemCapabilities,
+            builtInMetricsStore
+        )
     }
 
     private fun onBortEnabled(intent: Intent) {